requirements.txt
# ================================
# 数据要塞项目依赖
# 来自夜的命名术·壹的数字堡垒
# ================================

# 核心依赖
PyYAML>=6.0
types-PyYAML>=6.0.0  # PyYAML类型存根，用于mypy类型检查
cryptography>=3.4.8
psutil>=5.8.0

# 网络和通信
requests>=2.28.0
websockets>=10.0

# 系统监控
prometheus-client>=0.15.0
psycopg2-binary>=2.9.0

# 日志和监控
loguru>=0.6.0
rich>=12.0.0

# 开发和测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=5.0.0
black>=22.0.0
mypy>=0.991

# 安全相关
bcrypt>=4.0.0
argon2-cffi>=21.0.0